+++ public/index.html
@@ -19,7 +19,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>LanGauge</title>
   </head>
   <body>
     <noscript>
@@ -37,7 +37,7 @@
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
     <script src="https://www.gstatic.com/firebasejs/4.3.0/firebase.js"></script>
-    
+
     <!-- Latest compiled and minified CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
