--- src/Home.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react'
-import './App.css'
-
-class Home extends Component {
-  render () {
-    return (
-      <div>
-        <h1>Home</h1>
-      </div>
-    )
-  }
-}
-
-export default Home
