+++ src/homepage/Home.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react'
+import '../App.css'
+
+
+
+
+export default class Home extends Component {
+  render () {
+    return (
+      <div>
+        <h1>Home</h1>
+
+      </div>
+    )
+  }
+}
